[refactor] remove `EngineGraphPartitioner` and rename `SpecialTypeGraphPartitioner` to `EngineGraphPartitioner`
